Add support for default values.
Default values can be indicated with a "dict"-based config of the columns. Something like this:

    ORM("...", {
        "table": "my_table",
        "columns": {
            "foo": { "type": int, "default" 5 }
        }
    })

The old syntax (where column values were just the types) is still
supported.

Columns are not required to have a default value. If a default value
is not provided, a value of "None" is assumed (irrespective of the
type of the column). When an object is mapped, *ALL* of the columns
will be present. Those that are not supplied with keyword arguments
will have the default value if one is supplied, or None otherwise. For
example:

  Person = ORM(":memory:", {
      "table": "people",
      "columns": {
         "first_name": str,
         "last_name": { "type": str, "default": "Doe" }
      }
  }

  john_doe = Person(first_name="John", last_name="Doe")
  jane_doe = Person(first_name="Jane")
  sam_doe = Person(first_name="Sam", age=20)

All of the above will have the last name "Doe". john_doe.age and
jane_doe.age will return None, while sam_doe.age will return 20.